refactor(agentTool): Extract executeEngineAction as inverse of createEngineRequests
Moves the tool-invoke + ITaskData-wrap logic out of resolveSubAgentRequest into utils/agent-execution/executeEngineAction.ts, sitting next to createEngineRequests as its symmetric pair. resolveSubAgentRequest shrinks from 122 to 64 lines and only contains sub-agent loop control + HITL guard. Also trims verbose JSDoc and UserError description per code-comment rule.
