+++ packages/@n8n/nodes-langchain/utils/agent-execution/index.ts
@@ -9,6 +9,7 @@
  */
 
 export { createEngineRequests } from './createEngineRequests';
+export { executeEngineAction } from './executeEngineAction';
 export { buildResponseMetadata } from './buildResponseMetadata';
 export { buildSteps } from './buildSteps';
 export { processEventStream } from './processEventStream';
